Clean up; fix build
